refactor: extract CustomNodeLabel into separate component
Extract the CustomNodeLabel component from CustomNode.tsx into its own
file (CustomNodeLabel.tsx) to improve code organization and maintainability.

Benefits:
- Separation of concerns: Node rendering vs label editing logic
- Reduced file size: CustomNode.tsx (374 lines) + CustomNodeLabel.tsx (161 lines)
- Better maintainability: Easier to find and modify label-specific code
- Improved testability: Can test label behavior in isolation if needed
- Clearer responsibilities:
  * CustomNode: positioning, toolbar, drag/drop, node appearance
  * CustomNodeLabel: text display, editing, validation, state management

What moved to CustomNodeLabel.tsx:
- All label-related state (isEditing, editValue, isCancelling)
- All label handlers (handleDoubleClick, handleSave, handleCancel, etc.)
- Label rendering logic (span vs TextInput)
- Dynamic sizing calculations
- Props interface (CustomNodeLabelProps)
- Required imports (Icon, TextInput, ExclamationCircleIcon, etc.)

What was removed from CustomNode.tsx:
- Unused NodeLabelType import (now only in CustomNodeLabel.tsx)
- Unused React imports (clsx, useState, useEffect)

Test strategy:
- Kept existing tests in CustomNode.test.tsx (integration testing)
- All 8 tests continue to pass
- Tests verify label behavior in the context of the full node
- Can add CustomNodeLabel.test.tsx later for isolated unit tests if needed

Addresses Priority 3 from PR review #3085
