+++ packages/ui/src/components/Visualization/Custom/Node/CustomNode.tsx
@@ -1,7 +1,5 @@
 import './CustomNode.scss';
 
-import { Icon, TextInput } from '@patternfly/react-core';
-import { ExclamationCircleIcon } from '@patternfly/react-icons';
 import {
   AnchorEnd,
   DEFAULT_LAYER,
@@ -27,14 +25,12 @@ import {
   withContextMenu,
   withSelection,
 } from '@patternfly/react-topology';
-import clsx from 'clsx';
-import { FunctionComponent, useCallback, useContext, useEffect, useMemo, useRef, useState } from 'react';
+import { FunctionComponent, useCallback, useContext, useMemo, useRef } from 'react';
 
 import { CatalogModalContext } from '../../../../dynamic-catalog/catalog-modal.provider';
 import { useProcessorIcon } from '../../../../hooks/processor-icon.hook';
 import { useEntityContext } from '../../../../hooks/useEntityContext/useEntityContext';
 import { AddStepMode, IVisualizationNode, NodeToolbarTrigger } from '../../../../models';
-import { NodeLabelType } from '../../../../models/settings';
 import { CamelRouteVisualEntityData } from '../../../../models/visualization/flows/support/camel-component-types';
 import { SettingsContext } from '../../../../providers';
 import { NodeInteractionAddonContext } from '../../../registers/interactions/node-interaction-addon.provider';
@@ -45,6 +41,7 @@ import { NodeContextMenuFn } from '../ContextMenu/NodeContextMenu';
 import { getDropTargetContainerClassNames, GROUP_DRAG_TYPE, NODE_DRAG_TYPE } from '../customComponentUtils';
 import { TargetAnchor } from '../target-anchor';
 import { CustomNodeContainer } from './CustomNodeContainer';
+import { CustomNodeLabel } from './CustomNodeLabel';
 import { checkNodeDropCompatibility, getNodeDragAndDropDirection, handleValidNodeDrop } from './CustomNodeUtils';
 
 type DefaultNodeProps = Parameters<typeof DefaultNode>[0];
@@ -54,156 +51,6 @@ interface CustomNodeProps extends DefaultNodeProps {
   /** Toggle node collapse / expand */
   onCollapseToggle?: () => void;
 }
-
-interface CustomNodeLabelProps {
-  label: string;
-  doesHaveWarnings?: boolean;
-  validationText?: string;
-  x?: number;
-  y?: number;
-  transform?: string;
-  width?: number;
-  height?: number;
-  className?: string;
-  vizNode?: IVisualizationNode;
-  onDescriptionChange?: (description: string) => void;
-  nodeLabelType?: NodeLabelType;
-}
-
-const CustomNodeLabel: FunctionComponent<CustomNodeLabelProps> = ({
-  label,
-  doesHaveWarnings = false,
-  validationText,
-  x,
-  y,
-  transform,
-  width = CanvasDefaults.DEFAULT_LABEL_WIDTH,
-  height = CanvasDefaults.DEFAULT_LABEL_HEIGHT,
-  className = 'custom-node__label',
-  vizNode,
-  onDescriptionChange,
-  nodeLabelType = NodeLabelType.Description,
-}) => {
-  const [isEditing, setIsEditing] = useState(false);
-  const [editValue, setEditValue] = useState('');
-  const [isCancelling, setIsCancelling] = useState(false);
-  const inputRef = useRef<HTMLInputElement>(null);
-  const nodeDefinition = vizNode?.getNodeDefinition();
-  const currentDescription = nodeDefinition?.description || '';
-
-  useEffect(() => {
-    if (isEditing && inputRef.current) {
-      inputRef.current.focus();
-      inputRef.current.select();
-    }
-  }, [isEditing]);
-
-  const handleDoubleClick = useCallback(
-    (event: React.MouseEvent) => {
-      // Only allow editing if:
-      // 1. vizNode and onDescriptionChange are available
-      // 2. The current node label setting is set to 'description'
-      if (!vizNode || !onDescriptionChange || nodeLabelType !== NodeLabelType.Description) {
-        return;
-      }
-      event.stopPropagation();
-      setEditValue(currentDescription);
-      setIsEditing(true);
-      setIsCancelling(false);
-    },
-    [currentDescription, vizNode, onDescriptionChange, nodeLabelType],
-  );
-
-  const handleSave = useCallback(() => {
-    const trimmedValue = editValue.trim();
-    if (onDescriptionChange && trimmedValue !== currentDescription) {
-      onDescriptionChange(trimmedValue);
-    }
-    setIsEditing(false);
-    setIsCancelling(false);
-  }, [editValue, currentDescription, onDescriptionChange]);
-
-  const handleCancel = useCallback(() => {
-    setIsCancelling(true);
-    setIsEditing(false);
-    setEditValue('');
-  }, []);
-
-  const handleKeyDown = useCallback(
-    (event: React.KeyboardEvent<HTMLInputElement>) => {
-      if (event.key === 'Enter') {
-        event.stopPropagation();
-        handleSave();
-      } else if (event.key === 'Escape') {
-        event.stopPropagation();
-        handleCancel();
-      }
-    },
-    [handleSave, handleCancel],
-  );
-
-  const handleBlur = useCallback(() => {
-    // Don't save if user is cancelling with Escape
-    if (!isCancelling) {
-      handleSave();
-    }
-  }, [handleSave, isCancelling]);
-
-  // When editing, expand the input box to provide more space for typing
-  // Normal: 180px width x 24px height
-  // Editing: 240px width x 40px height (accommodates PatternFly TextInput padding)
-  const editWidth = 240;
-  const editHeight = 40;
-
-  // Calculate x-position adjustment to keep the input centered when width changes
-  // Example: if width increases from 180px to 240px (60px difference),
-  // shift left by 30px (half the difference) to maintain center alignment
-  const widthDiff = editWidth - width;
-  const adjustedX = isEditing && x !== undefined ? x - widthDiff / 2 : x;
-
-  return (
-    <foreignObject
-      width={isEditing ? editWidth : width}
-      height={isEditing ? editHeight : height}
-      className={className}
-      {...(transform ? { transform } : { x: adjustedX!, y: y! })}
-    >
-      <div
-        className={clsx('custom-node__label__text', {
-          'custom-node__label__text__error': doesHaveWarnings,
-        })}
-      >
-        {doesHaveWarnings && (
-          <Icon status="danger" title={validationText} data-warning={doesHaveWarnings}>
-            <ExclamationCircleIcon />
-          </Icon>
-        )}
-        {isEditing ? (
-          <TextInput
-            ref={inputRef}
-            type="text"
-            value={editValue}
-            onChange={(_event, value) => setEditValue(value)}
-            onKeyDown={handleKeyDown}
-            onBlur={handleBlur}
-            placeholder="Enter description"
-            aria-label="Edit node description"
-            onClick={(e) => e.stopPropagation()}
-          />
-        ) : (
-          <span
-            title={label}
-            onClick={(e) => e.stopPropagation()}
-            onDoubleClick={handleDoubleClick}
-            style={{ cursor: nodeLabelType === NodeLabelType.Description ? 'pointer' : 'default' }}
-          >
-            {label}
-          </span>
-        )}
-      </div>
-    </foreignObject>
-  );
-};
 
 function getShouldShowToolbar(
   trigger: NodeToolbarTrigger | undefined,
